# ISPN-9263 Segmented Off Heap Data Container