--- a/core/src/main/java/org/infinispan/container/offheap/MemoryAddressHash.java
+++ b/core/src/main/java/org/infinispan/container/offheap/MemoryAddressHash.java
@@ -2,6 +2,8 @@ package org.infinispan.container.offheap;
 
 import java.util.stream.LongStream;
 
+import org.infinispan.commons.util.Util;
+
 import sun.misc.Unsafe;
 
 /**
@@ -11,16 +13,16 @@ import sun.misc.Unsafe;
 public class MemoryAddressHash {
    private static final Unsafe UNSAFE = UnsafeHolder.UNSAFE;
    private static final OffHeapMemory MEMORY = OffHeapMemory.INSTANCE;
-   private static final int MAXIMUM_CAPACITY = 1 << 30;
-   private static final int HASH_BITS = 0x7fffffff; // usable bits of normal node hash
 
    private final long memory;
    private final int pointerCount;
+   private final OffsetCalculator offSetCalculator;
    private final OffHeapMemoryAllocator allocator;
 
-   public MemoryAddressHash(int pointers, OffHeapMemoryAllocator allocator) {
-      this.pointerCount = nextPowerOfTwo(pointers);
+   public MemoryAddressHash(int pointers, OffsetCalculator offSetCalculator, OffHeapMemoryAllocator allocator) {
+      this.pointerCount = Util.findNextHighestPowerOfTwo(pointers);
       long bytes = ((long) pointerCount) << 3;
+      this.offSetCalculator = offSetCalculator;
       this.allocator = allocator;
       memory = allocator.allocate(bytes);
       // Have to clear out bytes to make sure no bad stuff was read in
@@ -28,9 +30,7 @@ public class MemoryAddressHash {
    }
 
    private long findOffset(Object instance) {
-      int h = spread(instance.hashCode());
-      int pointerMask = pointerCount - 1;
-      return h & pointerMask;
+      return offSetCalculator.calculateOffset(instance);
    }
 
    public void putMemoryAddress(Object instance, long address) {
@@ -74,19 +74,5 @@ public class MemoryAddressHash {
             .limit(pointerCount)
             .map(l -> UNSAFE.getAndSetLong(null, l, 0))
             .filter(l -> l != 0);
-   }
-
-   private static final int nextPowerOfTwo(int c) {
-      int n = c - 1;
-      n |= n >>> 1;
-      n |= n >>> 2;
-      n |= n >>> 4;
-      n |= n >>> 8;
-      n |= n >>> 16;
-      return (n < 0) ? 1 : (n >= MAXIMUM_CAPACITY) ? MAXIMUM_CAPACITY : n + 1;
-   }
-
-   static final int spread(int h) {
-      return (h ^ (h >>> 16)) & HASH_BITS;
    }
 }