… core/src/main/java/org/infinispan/container/offheap/BoundedOffHeapDataContainer.java
… core/src/main/java/org/infinispan/container/offheap/OffHeapConcurrentMap.java
… core/src/main/java/org/infinispan/container/offheap/OffHeapDataContainer.java
… core/src/main/java/org/infinispan/container/offheap/SegmentedBoundedOffHeapDataContainer.java
… core/src/main/java/org/infinispan/factories/DataContainerFactory.java
… core/src/test/java/org/infinispan/container/offheap/OffHeapSingleNodeTest.java
… core/src/test/java/org/infinispan/stats/SingleStatsTest.java
… server/rest/src/test/java/org/infinispan/rest/BaseRestOperationsTest.java
… server/rest/src/test/java/org/infinispan/rest/assertion/ResponseAssertion.java