* Add off heap concurrent map to be used by segmented container

https://issues.jboss.org/browse/ISPN-9263
https://issues.jboss.org/browse/ISPN-8380