@wburns 2018-08-24
There are some related CI failures, but I was unable to reproduce them locally. I will have to keep trying.

https://ci.infinispan.org/job/Infinispan/job/PR-6217/3/#showFailuresLink