@wburns 2018-08-24
Think I fixed the ExceptionEvictionTest, it wasn't using the underlying storage key to do the peek. Not sure why it didn't fail locally.

Also the rest off heap appears to assume the entries are returned in a specific order, however there is no guarantee in which order the entries would be returned. Will have to make the test a bit more resilient to this.