@wburns 2018-08-24
Tests all pass now. Just need to clean up code a little bit.