@wburns 2018-08-27
CI passed with https://ci.infinispan.org/view/Infinispan/job/Infinispan/view/change-requests/job/PR-6217/5/ I refactored the entryStream a bit to be more OOM friendly, should still be fine though.