@wburns 2018-08-28
This also fixed https://issues.jboss.org/browse/ISPN-8380 since the new off heap containers use composition instead and remove expired entries. Tests have been updated to reflect that for off heap.