@wburns 2018-08-28
Performance for iteration/size results are pretty good. I am especially surprised that iteration is that much faster, as it still has to overcome the network overhead for returning all entries.

This is using https://github.com/infinispan/infinispan-benchmarks/tree/master/iteration configured to just use off heap with default settings with the following command

java -jar benchmark.jar -pvalueObjectSize=1024 -pkeyObjectSize=64 -pentryAmount=100000 -pbatchSize=16384 -puseIdentityCache=false

MASTER

Benchmark | (batchSize) | (entryAmount) | (keyObjectSize) | (nodes) | (useIdentityCache) | (useStrings) | (valueObjectSize) | Mode | Cnt | Score |   | Error | Units
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
JMHBenchmarks.iteratorParallel | 16384 | 100000 | 64 | 6 | FALSE | FALSE | 1024 | thrpt | 20 | 2.932 | ± | 0.073 | ops/s
JMHBenchmarks.iteratorSequential | 16384 | 100000 | 64 | 6 | FALSE | FALSE | 1024 | thrpt | 20 | 1.405 | ± | 0.034 | ops/s
JMHBenchmarks.noRehashIteratorParallel | 16384 | 100000 | 64 | 6 | FALSE | FALSE | 1024 | thrpt | 20 | 3.154 | ± | 0.125 | ops/s
JMHBenchmarks.noRehashIteratorSequential | 16384 | 100000 | 64 | 6 | FALSE | FALSE | 1024 | thrpt | 20 | 1.65 | ± | 0.03 | ops/s
JMHBenchmarks.sizeParallel | 16384 | 100000 | 64 | 6 | FALSE | FALSE | 1024 | thrpt | 20 | 7.698 | ± | 0.299 | ops/s
JMHBenchmarks.sizeSequential | 16384 | 100000 | 64 | 6 | FALSE | FALSE | 1024 | thrpt | 20 | 2.406 | ± | 0.569 | ops/s

NEW

Benchmark | (batchSize) | (entryAmount) | (keyObjectSize) | (nodes) | (useIdentityCache) | (useStrings) | (valueObjectSize) | Mode | Cnt | Score |   | Error | Units
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
JMHBenchmarks.iteratorParallel | 16384 | 100000 | 64 | 6 | FALSE | FALSE | 1024 | thrpt | 20 | 4.418 | ± | 0.093 | ops/s
JMHBenchmarks.iteratorSequential | 16384 | 100000 | 64 | 6 | FALSE | FALSE | 1024 | thrpt | 20 | 2.3 | ± | 0.056 | ops/s
JMHBenchmarks.noRehashIteratorParallel | 16384 | 100000 | 64 | 6 | FALSE | FALSE | 1024 | thrpt | 20 | 4.655 | ± | 0.307 | ops/s
JMHBenchmarks.noRehashIteratorSequential | 16384 | 100000 | 64 | 6 | FALSE | FALSE | 1024 | thrpt | 20 | 2.626 | ± | 0.042 | ops/s
JMHBenchmarks.sizeParallel | 16384 | 100000 | 64 | 6 | FALSE | FALSE | 1024 | thrpt | 20 | 41.527 | ± | 0.45 | ops/s
JMHBenchmarks.sizeSequential | 16384 | 100000 | 64 | 6 | FALSE | FALSE | 1024 | thrpt | 20 | 18.951 | ± | 0.119 | ops/s

Performance Increase

Benchmark | Percent Increase
-- | --
JMHBenchmarks.iteratorParallel | 50.68%
JMHBenchmarks.iteratorSequential | 63.70%
JMHBenchmarks.noRehashIteratorParallel | 47.59%
JMHBenchmarks.noRehashIteratorSequential | 59.15%
JMHBenchmarks.sizeParallel | 439.45%
JMHBenchmarks.sizeSequential | 687.66%