@wburns 2018-08-28
Can confirm that bounded off heap iteration has the same perf essentially as unbounded. The difference will be in read/write perf which I will test to make sure the new bounded container doesn't have a regression there.