@wburns 2018-08-28
Testing put and get

Unbounded is comparable to master, with gets being a bit faster.

UNBOUNDED

Benchmark | Percent Increase
-- | --
JMHBenchmarks.iteratorParallel | 3.36%
JMHBenchmarks.iteratorSequential | 3.71%
JMHBenchmarks.noRehashIteratorParallel | 0.14%
JMHBenchmarks.noRehashIteratorSequential | 15.86%
JMHBenchmarks.sizeParallel | 1.07%

However bounded is a bit slower for all ops

BOUNDED

Benchmark | Percent Increase
-- | --
JMHBenchmarks.iteratorParallel | -3.20%
JMHBenchmarks.iteratorSequential | -3.06%
JMHBenchmarks.noRehashIteratorParallel | -4.62%
JMHBenchmarks.noRehashIteratorSequential | -5.20%
JMHBenchmarks.sizeParallel | -4.28%

I will try to see if there is anything I can do to speed the bounded variants, but I am not optimistic.