@wburns 2018-08-28
I ran it a second time and the numbers changed quite a bit, both were faster and under heavier concurrency the new bounded segmented container was faster, however individual read/writes were slower.


Benchmark | Percent Difference
-- | --
JMHBenchmarks.iteratorParallel | 2.65%
JMHBenchmarks.iteratorSequential | 2.73%
JMHBenchmarks.noRehashIteratorParallel | 1.91%
JMHBenchmarks.noRehashIteratorSequential | -8.34%
JMHBenchmarks.sizeParallel | -4.62%

I will try to see if I can find any way to improve performance more, but no changes so far have yielded anything and the code is very similar to before.