@wburns 2018-08-30
Added in replace implementations as well.