@wburns 2018-09-06
@danberindei these test failures are the ones related to jgroups update ? 
https://ci.infinispan.org/job/Infinispan/job/PR-6217/9/#showFailuresLink

Just curious because my PR had no failures before rebase to master.