@wburns 2018-09-07
Pretty sure the jgroups update caused the failures on CI, rebasing just to make sure they pass now though.